// src/app/models/defineClass.ts
export class Student {
    // This is for the object id auto generated by mongo
    _id?: string;
    FirstName: string;
    LastName: string;
    School: string;
    StartDate: Date;
}

export class Card {
    // This is for the object id auto generated by mongo
    _id?:       string;
    OwnerName:  string;
    CardName:   string;
    CardAddress: string;
    Password:   string;
    StartDate:  Date;
}

export class CardBalance {
    balance: string;
    nonce:   string;
    type:    number;
}

export class NetState {

    public chain_id: number
    public tail: string;
    public lib: string;
    public height: string;
    public protocol_version: string;
    public synchronized: boolean;
    public version: string
}

export class PwdChangeStatusModel {
    _body:string;
    status:number;
    statusText:string;
}
  

export class NetStatusResult {
    chain_id: number;
    height: string;
    lib: string;
    protocol_version: string;
    synchronized: boolean;
    tail: string;
    version: string;
}

export class NetStatusModel {
    result: NetStatusResult;
}


export class signedTransaction {
    nonce: number;
    file: any;
}